scrapped version 6 and restarting
