+++ client/app/AppRoutes.tsx
@@ -2,9 +2,7 @@ import React from "react";
 import crAppTheme from "./theme";
 import { ThemeProvider } from "@mui/material";
 import { Route, Routes } from "react-router-dom";
-import Home from "../features/home/Home";
 import StyleGuide from "../features/styles/StyleGuide";
-import Profile from "../features/navigation/Profile";
 
 /**
  * COMPONENT
@@ -14,10 +12,7 @@ const AppRoutes = () => {
   return (
     <ThemeProvider theme={crAppTheme}>
       <Routes>
-        {/* <Route path="/*" element={<Home />} />
-        <Route to="/home" element={<Home />} /> */}
         <Route path="/styleguide" element={<StyleGuide />} />
-        <Route path="/profile" element={<Profile />} />
       </Routes>
     </ThemeProvider>
   );
